finregistry-pipeline: Reorganize
- Move functions related to prepping the cohort and cases to survival_analysis.py and keep the functions related to sampling cases and controls in sample.py

- Move key figures, distributions and cumulative incidence from summary_stats.py to their own functions

- Update the pipeline script
