+++ pipeline/scripts/pipeline.py
@@ -1,20 +1,24 @@
 """FinRegistry pipeline"""
 
+from risteys_pipeline.config import FINREGISTRY_OUTPUT_DIR
 from risteys_pipeline.finregistry.load_data import (
     load_endpoints_data,
     load_minimal_phenotype_data,
     load_first_events_data,
 )
-from risteys_pipeline.finregistry.summary_stats import compute_distribution, compute_key_figures, cumulative_incidence
-from risteys_pipeline.finregistry.sample import get_cohort
-from risteys_pipeline.finregistry.survival_analysis import prep_all_cases
+from risteys_pipeline.finregistry.key_figures import compute_key_figures
+from risteys_pipeline.finregistry.distributions import compute_distribution
+from risteys_pipeline.finregistry.cumulative_incidence import cumulative_incidence
+from risteys_pipeline.finregistry.survival_analysis import get_cohort, prep_all_cases
+from risteys_pipeline.finregistry.write_data import (
+    get_output_filepath,
+    summary_stats_to_json,
+    write_json_to_file,
+)
+
 
 def pipeline():
     """Running the Risteys pipeline on FinRegistry data"""
-
-    # Set up output path
-    # TODO: move to config
-    output_path = "~/risteys/results"
 
     # Read data
     endpoints = load_endpoints_data()
@@ -25,21 +29,32 @@ def pipeline():
     cohort = get_cohort(minimal_phenotype)
     all_cases = prep_all_cases(first_events, cohort)
 
-    # Run key figures and write to a file
+    # Run key figures
     kf_all = compute_key_figures(first_events, minimal_phenotype, index_persons=False)
-    kf_all.to_csv(output_path + "/key_figures_all.csv", index=False)
-    kf_index_persons = compute_key_figures(first_events, minimal_phenotype, index_persons=True)
-    kf_index_persons.to_csv(output_path + "/key_figures_index_persons.csv", index=False)
+    kf_index_persons = compute_key_figures(
+        first_events, minimal_phenotype, index_persons=True
+    )
 
-    # Run age and year distributions and write to a file
+    # Run age and year distributions
     dist_age = compute_distribution(first_events, "age")
-    dist_age.to_csv(output_path + "/distribution_age.csv", index=True)
     dist_year = compute_distribution(first_events, "year")
-    dist_year.to_csv(output_path + "/distribution_year.csv", index=True)
+
+    # Write key figures and distributions to file
+    summary_stats_json = summary_stats_to_json(kf_all, dist_age, dist_year)
+    kf_json = kf_index_persons.set_index("endpoint").to_json(orient="index")
+    write_json_to_file(
+        summary_stats_json,
+        get_output_filepath("summary_stats", "json", FINREGISTRY_OUTPUT_DIR),
+    )
+    write_json_to_file(
+        kf_json,
+        get_output_filepath("key_figures_index", "json", FINREGISTRY_OUTPUT_DIR),
+    )
 
     # Run cumulative incidence and write to a file
     ci = cumulative_incidence(cohort, all_cases, endpoints)
-    ci.to_csv(output_path + "/cumulative_incidence.csv", index=False)
+    ci.to_csv(FINREGISTRY_OUTPUT_DIR / "cumulative_incidence.csv", index=False)
+
 
 if __name__ == "__main__":
     pipeline()
